(maint) add support for additional files, update sort functions
This commit refactors the project. Instead of hardcoding the three test files,
it now accepts any number of file paths as arguments. It then parses the first
line of the file to determine which type of file it is.

Additionally, the merge-sort functions were replaced with simpler calls
to clojure's built in sorting method. The old methods remain for demonstration
purposes, but they are no longer used.

Tests continue to pass.